Refactoring: lets register our commands via init()
This simplifies command creation for test purposes
